requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
psycopg2-binary==2.9.1
python-dotenv==0.19.0